Control line spacing, and debounce text.
I wanted the text to only redraw when the user isn't actively typing,
so I added a DispatcherTimer to the class, and use it to only paint
when the user hasn't typed for 1 second.

Also, I added Line Spacing to the list of UI-settable parameters. The
implementation parallels that of Character Spacing.